services/intelligence-service/requirements.txt
fastapi
uvicorn
pydantic
redis
google-genai
python-dotenv